+++ lib/src/search/base_search_content.dart
@@ -1,2 +1,6 @@
-/// This can either be a [SearchVideo] or [SearchPlaylist]
+import 'search_channel.dart';
+import 'search_playlist.dart';
+import 'search_video.dart';
+
+/// This can either be a [SearchVideo], [SearchPlaylist], [SearchChannel]
 mixin BaseSearchContent {}