+++ lib/src/search/search_query.dart
@@ -1,5 +1,5 @@
+import '../../youtube_explode_dart.dart';
 import '../reverse_engineering/pages/search_page.dart';
-import '../reverse_engineering/youtube_http_client.dart';
 
 ///
 class SearchQuery {
@@ -15,8 +15,9 @@ class SearchQuery {
 
   /// Search a video.
   static Future<SearchQuery> search(
-      YoutubeHttpClient httpClient, String searchQuery) async {
-    var page = await SearchPage.get(httpClient, searchQuery);
+      YoutubeHttpClient httpClient, String searchQuery,
+      {SearchFilter filter = const SearchFilter('')}) async {
+    var page = await SearchPage.get(httpClient, searchQuery, filter: filter);
     return SearchQuery(httpClient, searchQuery, page);
   }
 
@@ -31,7 +32,7 @@ class SearchQuery {
   }
 
   /// Content of this search.
-  /// Contains either [SearchVideo] or [SearchPlaylist]
+  /// Contains either [SearchVideo], [SearchPlaylist] or [SearchChannel]
   List<dynamic> get content => _page.searchContent;
 
   /// Videos related to this search.