+++ lib/src/videos/streams/streams.dart
@@ -6,7 +6,7 @@ export 'framerate.dart';
 export 'muxed_stream_info.dart';
 export 'stream_container.dart';
 export 'stream_context.dart';
-export 'stream_info.dart' hide mediaTypeFromJson, mediaTypeTojson;
+export 'stream_info.dart' hide mediaTypeFromJson, mediaTypeToJson;
 export 'stream_manifest.dart';
 export 'streams_client.dart';
 export 'video_only_stream_info.dart';