+++ test/search_test.dart
@@ -1,5 +1,10 @@
+import 'dart:math';
+
 import 'package:test/test.dart';
 import 'package:youtube_explode_dart/youtube_explode_dart.dart';
+
+final rnd = Random();
+const letters = r'abcdefghilmnopqrstuvzjkwy1234567890!@#$%^&*()_+{}|"?><|~`|';
 
 void main() {
   YoutubeExplode? yt;
@@ -12,47 +17,64 @@ void main() {
   });
 
   test('Search a youtube video from the search page', () async {
-    var videos = await yt!.search.getVideos('undead corporation megalomania');
+    var videos = await yt!.search.search('undead corporation megalomania');
     expect(videos, isNotEmpty);
   });
 
-  test('Search a youtube video from the search page-2', () async {
-    var videos = await yt!.search
-        // ignore: deprecated_member_use_from_same_package
-        .getVideosFromPage('hello')
-        .where((e) => e is SearchVideo) // Take only the videos.
-        .cast<SearchVideo>()
-        .take(10) // Take on 10 results.
-        .toList();
-    expect(videos, hasLength(10));
-    var video = videos.first;
-    expect(video.id, isNotNull);
-
-    expect(video.title, isNotEmpty);
-    expect(video.author, isNotEmpty);
-    expect(video.description, isNotEmpty);
-    expect(video.duration, isNotEmpty);
-    expect(video.viewCount, greaterThan(0));
-    expect(video.thumbnails, isNotEmpty);
+  test('Search with no results', () async {
+    var videos = await yt!.search.search(
+        List.generate(1300, (_) => letters[rnd.nextInt(letters.length)])
+            .join());
+    expect(videos, isEmpty);
+    var nextPage = await videos.nextPage();
+    expect(nextPage, isNull);
   });
 
-  // Seems all search key also have result now, so hide this test case
-  // test('Search with no results - old', () async {
-  //   var query =
-  //       // ignore: deprecated_member_use_from_same_package
-  //       await yt!.search.queryFromPage('g;jghEOGHJeguEPOUIhjegoUEHGOGHPSASG');
-  //   expect(query.content, isEmpty);
-  //   expect(query.relatedVideos, isEmpty);
-  //   var nextPage = await query.nextPage();
-  //   expect(nextPage, isNull);
-  // });
+  test('Search only channels', () async {
+    var channels = await yt!.search
+        .searchContent('PewDiePie', filter: TypeFilters.channel);
+    expect(channels.every((e) => e is SearchChannel), isTrue);
+  });
 
-  test('Search youtube videos have thumbnails - old', () async {
-    // ignore: deprecated_member_use_from_same_package
-    var searchQuery = await yt!.search.queryFromPage('hello');
-    expect(searchQuery.content.first, isA<SearchVideo>());
+  test('Search only playlists', () async {
+    var channels =
+        await yt!.search.searchContent('Banana', filter: TypeFilters.playlist);
+    expect(channels.every((e) => e is SearchPlaylist), isTrue);
+  });
 
-    var video = searchQuery.content.first as SearchVideo;
-    expect(video.thumbnails, isNotEmpty);
+  test('Search only movies', () async {
+    var channels =
+        await yt!.search.searchContent('Banana', filter: TypeFilters.playlist);
+    expect(channels.every((e) => e is SearchPlaylist), isTrue);
+  });
+
+  test('Search test search filters', () async {
+    var featureSearch =
+        await yt!.search.searchContent('hello', filter: FeatureFilters.hd);
+    expect(featureSearch, isNotEmpty);
+
+    var uploadSearch = await yt!.search
+        .searchContent('hello', filter: UploadDateFilter.lastHour);
+    expect(uploadSearch, isNotEmpty);
+
+    var durationSearch =
+        await yt!.search.searchContent('hello', filter: DurationFilters.long);
+    expect(durationSearch, isNotEmpty);
+
+    var sortSearch =
+        await yt!.search.searchContent('hello', filter: SortFilters.viewCount);
+    expect(sortSearch, isNotEmpty);
+  });
+
+  test('Search raw', () async {
+    var search = await yt!.search.searchRaw('hello');
+    expect(search.content, isNotEmpty);
+    expect(search.relatedVideos, isNotEmpty);
+    expect(search.estimatedResults, greaterThan(1));
+  });
+
+  test('Get youtube search suggestions', () async {
+    var suggestions = await yt!.search.getQuerySuggestions('hello');
+    expect(suggestions, isNotEmpty);
   });
 }